requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-unfold==0.64.2
idna==3.10
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
stripe==12.4.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
